chore: add language button to login page
