+++ apps/frontend/.eslintrc.js
@@ -1,4 +1,9 @@
 module.exports = {
+  parserOptions: {
+    project: 'tsconfig.json',
+    tsconfigRootDir: __dirname,
+    sourceType: 'module',
+  },
   root: true,
   extends: ['custom/next'],
 };
