+++ apps/frontend/components/pages/Login/Login.tsx
@@ -5,6 +5,7 @@ import { NextPage } from 'next/types';
 import { useForm } from 'react-hook-form';
 import { FormattedMessage, useIntl } from 'react-intl';
 
+import SelectLanguage from 'components/SelectLanguage';
 import { Input, PasswordInput } from 'components/atoms';
 import { Pages } from 'constant';
 import { login, LoginData } from 'services/api/auth/login';
@@ -98,6 +99,7 @@ export const Login: NextPage = () => {
             <FormattedMessage id="login.submit" />
           </button>
         </form>
+        <SelectLanguage />
       </div>
     </main>
   );
